Skip request permission except on Android
