+++ example/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import { PermissionsAndroid, StyleSheet, View, Button } from 'react-native';
+import { Platform, StyleSheet, View, Button } from 'react-native';
 
+import usePermission from './usePermission';
 import RecorderExample from './RecorderExample';
 import PlayerExample from './PlayerExample';
 
@@ -15,18 +16,7 @@ const styles = StyleSheet.create({
 });
 
 export default function App() {
-  const [permissionGranted, setPermissionGranted] = React.useState();
-
-  const requestPermission = React.useCallback(async () => {
-    try {
-      const result = await PermissionsAndroid.request(
-        PermissionsAndroid.PERMISSIONS.RECORD_AUDIO,
-      );
-      setPermissionGranted(result === PermissionsAndroid.RESULTS.GRANTED);
-    } catch (err) {
-      console.warn(err);
-    }
-  }, []);
+  const [permissionGranted, requestPermission] = usePermission();
 
   React.useEffect(() => {
     requestPermission();
@@ -34,15 +24,17 @@ export default function App() {
 
   return (
     <View style={styles.root}>
-      <Button
-        title={
-          permissionGranted
-            ? 'Record Audio Permission Granted!'
-            : 'Request Record Audio Permission'
-        }
-        onPress={requestPermission}
-        disabled={permissionGranted}
-      />
+      {Platform.OS === 'android' && (
+        <Button
+          title={
+            permissionGranted
+              ? 'Record Audio Permission Granted!'
+              : 'Request Record Audio Permission'
+          }
+          onPress={requestPermission}
+          disabled={permissionGranted}
+        />
+      )}
 
       <View style={styles.container}>
         <RecorderExample disabled={!permissionGranted} />
